python/tag.cc: Fix invalid read in TagFileNext
We copied the string with snprintf("%s") which was wrong, because
it was not null-terminated. Explicitly copy all bytes with memcpy()
and then add the extra newline and nul terminator to the end.

(cherry picked from commit e9910d9c62a77d628f5278c47690b2f0b0a3fc3a)