apt/package.py: Use strongest avaialble hash when fetching
Use the strongest hash available when fetching binaries and
source packages. This requires a bit of trickery compared to
later releases where HashStringList is supported everywhere,
but it's not that bad.

We do not reject files with only untrusted hashes, placing
the burden of making the repository safe on the signer.

Closes: #944696
LP: #1858972
CVE-2019-15795